Logo changed. Added game type check. Ignoring pdf files

+++ LZRStatsApi/Models/ModelExtensions/GameExtensionMethods.cs
@@ -10,7 +10,7 @@ namespace LZRStatsApi.Models.ModelExtensions
     {
         public static bool IsPlayoffGame(this Game game)
         {
-            return game.GameType == (int)GameType.Playoff;
+            return game.GameType == GameType.Playoff;
         }
     }
 }
